排行榜改成时间

+++ BubbleKnight/Assets/BubbleKnight/Scripts/GameManager.cs
@@ -145,7 +145,7 @@ public class GameManager : MonoBehaviour
         if (state == GameState.Win)
         {
             AudioManager.Instance.PlaySound(5);
-            LocalLeaderboard.instance.SaveScore((int)nowTime);
+            LocalLeaderboard.instance.SaveScore(nowTime);
         }
         else
         {
